requirements.txt
numpy
pandas
openpyxl
snakemake >=7.0
ruamel.yaml
biopython >=1.73
click_help_colors
cookiecutter
